Check if path exists before trying to remove

+++ tests/testing_utils/utils.py
@@ -17,7 +17,9 @@ def write_process_stdout_to_file(process, filepath):
 
 
 def clean_existing_results():
-    shutil.rmtree(RESULTS_DIR)
-    shutil.rmtree(SANDBOXES_DIR)
+    if os.path.exists(RESULTS_DIR):
+        shutil.rmtree(RESULTS_DIR)
+    if os.path.exists(SANDBOXES_DIR):
+        shutil.rmtree(SANDBOXES_DIR)
     os.mkdir(RESULTS_DIR)
     os.mkdir(SANDBOXES_DIR)
